Db -> CouchDbConnectionFactory; commonize some data access code; use stale views.
The CouchDbConnectionFactory rename was done with the intention of
bringing a Redis connection into this.  I'm not quite sure if I'll do
that, though, so it may be reverted to a less unwieldy name.  We'll see.

Stale views are used to speed up retrieval time.  (It's not _that_
important that we always see the latest results.)